Debugged k_best_assign_mult_cost_matrices
Was passing in the number of targets as a parameter, but this
may be different for different cost matrices.  This caused
incorrect partitioning of nodes because the number of measurements
is used to decide whether two assignment matrices correspond to
different measurement associations and target deaths (they may
not due to filler entries in the assignment matrices).